Moves performance calculations into a separate module.
